+++ ui/src/app/modules/status/widgets/widgets.interfaces.ts
@@ -87,5 +87,4 @@ export interface Widget {
   }
   showNpmVersion?: boolean // update info widget
   dockerExpanded?: boolean // update info widget
-  hideNodeInfo?: boolean // update info widget
 }